requirements.txt
numpy==1.16.2
torch==0.4.1
pytest==4.2.1
tensorflow_gpu==1.12.0
